pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-duplex-zh"
version = "0.1.0"
description = "Chinese-first full duplex speech LLM stack built around Whisper Medium, Qwen 7B, CosyVoice2, and Voila-Tokenizer."
authors = [
    { name = "SpeakLLM Team" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchaudio>=2.2",
    "transformers>=4.39",
    "datasets>=2.18",
    "accelerate>=0.28",
    "peft>=0.10",
    "sentencepiece",
    "soundfile",
    "opencc-python-reimplemented",
    "pypinyin",
    "g2pc",
    "webrtcvad-wheels",
    "onnxruntime-gpu>=1.17",
    "protobuf",
    "pyyaml",
    "typer[all]>=0.9",
    "rich",
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "grpcio",
    "grpcio-tools",
    "numpy",
    "scipy",
    "librosa",
    "evaluate",
    "jiwer",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
